+++ packages/gamut/src/Markdown/libs/overrides/Iframe/index.tsx
@@ -1,6 +1,7 @@
 /* eslint-disable jsx-a11y/iframe-has-title */
 import React, { FunctionComponent, HTMLAttributes } from 'react';
 
+// eslint-disable-next-line gamut/no-css-standalone
 import styles from './styles.module.scss';
 
 export interface IframeProps extends HTMLAttributes<HTMLIFrameElement> {